requirements.txt
fastapi
uvicorn[standard]==0.22.0

numpy==1.24.4
pandas==2.0.3

requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

prophet

passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

python-multipart==0.0.6
